refactor: /config separado em abas — Empresa, Canais, Integrações, Vendas
A tela empilhava sete assuntos sem relação numa coluna só. Agora cada um mora
na sua área, e a página fica com o cabeçalho, o estado do perfil e o seletor.

- EmpresaTab   — Perfil da empresa + Brand Kit + Prévia
- CanaisTab    — WhatsappSection + CelularSection
- IntegracoesTab — IntegrationsSection + GoogleSection
- VendasTab    — Funil de entrada padrão
- ui.tsx       — Card e Inp, que agora têm mais de um dono

Reorganização pura: nenhum campo novo, nenhuma regra nova. O painel só monta
quando a aba abre, então abrir /config não dispara mais a rede das quatro
áreas de uma vez.

O único teste alterado é o do funil, que o procurava sem clicar em aba — ele
mudou de endereço, não de comportamento. Os 23 de WhatsappSection.test.tsx e
o resto dos 512 passam sem tocar em nada.
